fix window and logger
